Port NC to use trapperkeeper
This is currently referring to a SNAPSHOT version of trapperkeeper.

We'll want to minimize the amount of time that that is true for,
as SNAPSHOTs can be a bit of a pain to work with.  Will update
this PR with a released version of TK if we have one before this
is merged.