Refactor http handlers.
Define a parametrized liberator resource called `classifier-resource`
which takes a map with :get, :put, and :delete keys mapping to functions
that implement those actions. Since the only differences between the old
node, group, and class resources were in the action functions, all of
those can be replaced with this parametrized resource.

Replace all the identical :malformed? decisions with parse-if-body,
which ensures that the request body is non-empty before trying to parse
it. The previous :malformed? decision assumed that if the request had a
body, it was non-empty, which is not true in practice.

Modify malformed-or-parse to store malformed request bodies in the
context so they can be retrieved by the appropriate handler in order to
echo the body in the error message.