Upgrade to clojure.java.jdbc 0.3.0
Unfortunately, this depends on changes in migratus to make it also work
with jdbc 0.3.0. Until those changes have been merged in to migratus and
a new version has been released, this branch will only work if you have
the `jdbc-0.3.0` branch of `https://github.com/aperiodic/migratus.git`
as a lein checkout dependency.