(NC-102) Ensure that groups don't duplicate classes
Add a primary key to the group_classes table on the only two columns in
that table, which ensures that rows won't be duplicated and speeds up
querying by indexing the table.

When inserting a group's class references into the database, we iterate
through a set of the group's classes in order to avoid inserting
duplicate references. Since this is the only time we insert into that
table, this is sufficient to ensure uniqueness (so we won't catch
exceptions thrown by the database that may be caused by violation of the
primary key uniqueness constraint).

Add a test that checks that we can insert a group with duplicate class
references without throwing exceptions, and that after retrieving said
class the duplicates have been removed.