(NC-130) Add API docs.
Add documentation for the routes & allowed methods that currently exist;
namely creating and retrieving nodes, environments, and groups; creating
rules; and retrieving classes and node classifications.

It is possible currently to create classes through the API, but this
mainly exists to allow us to test the API. Once we can receive class
information directly from Puppet, we will only be using those class
definitions and that API route will disappear. Therefore, the
documentation does not cover the class creation operation.