(NC-113) Add listing routes & tests, allow deletes
The listing routes are the object routes excepting the last component
(which specifies the object), and they return a JSON array of all of the
extant objects of the appropriate type.

The delete method was not previously allowed for crud-resource routes,
despite crud-resource expecting a delete method to be supplied, so add
it to the list of allowed methods and add more http and acceptance tests
for deletion of things.