(NC-133) Synchronize class definitions from puppet
This adds the ability to retrieve class definitions from puppet and merge them
with the internal definitions. A new endpoint at /v1/update-classes accepts a
POST request with no body to trigger this.

This depends on features in Puppet 3.5, which is unreleased. In order to test
this, the integration test code now allows installing development builds of
puppet.