(NC-106) Add update-group storage method & tests.
Given a delta to an existing group (that must have a :name key whose
value matches the group to update), modifies the classes, class
parameters, and variables of the group as specified in the delta. Paths
mapped to nil by the delta are deleted, any other value is updated or
inserted as appropriate.

The update transaction requires the stricter "repeatable read" transaction
level, since it does multiple select-like statements that should all see the
same sets of rows. Since this transaction can fail if concurrent transactions
change the sets of rows these statements would operate on, the whole transaction
is retried up to three times on such failures before we give up.