(NC-106) Add route to update group with a POST.
Allow updating a group via HTTP. Updates are just like creation
requests, except instead of containing the whole group in the body, they
just contain a delta. Paths that lead to a nil value in the delta are
deleted, others are inserted/updated as appropriate. Only classes, class
parameters, and variables in a group delta are processed; all other
attributes are ignored.

Add documentation on doing an update, describing the semantics of
deleting things from the group by using null values.