NC-122 Add more rule operators.
Add regex match and numeric comparison operators. The regex operator
compiles the string directly to a java.util.regex.Pattern. The numeric
comparison operators

Modify the Rule schema to only accept rules with a :when field that
looks like a valid rule condition (i.e. a three-element vector with
a known operator and two strings, or a two-element vector that has
a "not" first and then a valid condition (which can itself be a further
negation).