(NC-107) Update environments of groups.
Just do a whole-sale update of the environment_name field in the group,
group_classes, and group_class_parameters tables. Foreign-key
constraints will blow up if the classes the group sets parameters for
don't exist in the new environment, or do exist but don't have all the
parameters that the group sets. In the future, this would be a good
occasion to refresh our view of puppet's classes for the target
environment.