(packaging) Update version determination for new packaging tools
Previously one could access "@" variables supplied by the packaging tools
without scoping them in any way. This changed, for sanity's sake, but alas,
this means some projects have to change with it. This commit also updates the
Rakefile to fall back to `git` and then a string in the case that the packaging
repo is not present.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>