(NC-136) Update to TK 0.3.2 so we can use new KS.
Updating KitchenSink brings in a new version of clojure.tools.cli with
incompatible semantics that broke TrapperKeeper, so update to a new TK
that uses the new clojure.tools.cli so we can use the new KS that has
the deep-merge utility.

Change the line that the acceptance test waits for to decide that the
server has started, since the new TK WebserverService has slightly
different output.