(maint) sql aggregation utils preserve row order.
Due to the use of group-by, the aggregate-* sql utils did not preserve
the order of the rows passed in if there were more than eight groups,
since this is the point at which clojure switches from using an ArrayMap
(which preserves insertion order) to a HashMap (which does not). Since
none of our tests exceeded this threshold, it was not noticed before.

Add ordered-group-by to the sql-utils ns, which is similar to group-by,
but adds the first index that a member of the group was seen at to the
beginning of the vector with the group members, and returns a seq of map
entries sorted by this index (rather than a map).

Augment the aggregate-column test so it has enough data to overflow the
ArrayMap size and use a HashMap, which will mess up ordering if we're
not careful.