(maint) Fix some integration test issues.
First off I just broke them after adding the parent field to groups and
not updating the tests to supply it.

Fix a race condition between our clojure webapp compiling and
registering with Jetty (which happens after Jetty begins to accept HTTP
requests) and our first integration test running that would cause the
tests to fail some of the time on my machine. The fix is to poll a real
classifier endpoint (specifically '/v1/environments') in the
`sleep_until_started` helper until it stops 404ing (indicating that
Jetty is routing requests to the classifier handler).