(NC-156) Add utilities for making map deltas.
The deltas represent the difference between two maps according to the
semantics of merge-and-clean, so they can be applied to the first map
using merge-and-clean in order to produce the second map.

This is used to implement the group-delta function, which is almost
identical but ensures the output always has a name, making it a valid
GroupDelta.