(NC-158) Mark class params deleted in synchronize.
Previously, only the class was marked deleted while the parameters were
untouched, which would lead to an error in create-parameter when it
tried to re-insert parameters that already existed.

As an additional precaution against this sort of bug, create-parameter
is now idempotent, in that it just updates the parameter if it already
exists in the database (previously it would only update if the parameter
existed *and* its delete field was `true`).