(NC-111) Add `expand-seq-params` utility.
This utility relieves the tedium of using sequential SQL parameters in
prepared statements (e.g. with an `IN` clause) by expanding each
sequential parameter's placeholder in the prepared statement SQL to be
an SQL list literal with the appropriate number of `?`s given the length
of the sequential parameter.

For example:
  ["SELECT * FROM kerjiggers WHERE name IN ?" ["bob" "carol" "eve"]]
becomes
  ["SELECT * FROM kerjiggers WHERE name IN (?,?,?)" "bob" "carol" "eve"]