(NC-111) Return annotated groups from endpoints.
Whenever a group or groups would be returned to the user in a 200
response, they are annotated first to mark deleted class or class
parameters that the group references.

This does not happen for groups that are returned with a 201 response,
but since this only happens when the user is creating the group with
a PUT, and all the classes & parameters of the group are checked before
it is created, it should not be an issue.

Update API docs for groups to describe the annotations and provide an
example of an annotated group.