(NC-188) Begin the Great UUID Reorganization.
Change the database to make the group ID the primary key, rather than
the name. Accordingly, change the foreign key links to refer to group ID
over group name.

Change the Group schema to require the :id key and make :parent a UUID.
Change the GroupDelta schema similarly, to requir the :id key, make
:name optional, and make :parent a UUID.

Change the group-delta utility for creating a delta between two groups
to always have the :id and ensure the :ids are equal.

Change the Rule schema to have :group-id as an optional key, rather than
:group-name.

THIS BREAKS ALL THE THINGS. If you are git-bisecting, you should skip
all the NC-188 commits from this one up until where the branch is merged
back in.

change Rule schema - fixup somewhere