(NC-188) HTTP API uses UUID as sole group key.
Rather than determining whether the group identifier passed in the URI
is a UUID or a group name, the HTTP handler assumes it's a UUID and
tries to parse it as such.

Since the ID is always present in the URI but may also be submitted in
the body, there's a new way for a group request to be malformed: if
these two IDs are different.

All the HTTP tests now pass again, however...

MOST OF THE THINGS ARE STILL BROKEN. The rules, classification, and
acceptance tests still mostly error out. If you are bisecting, you
should probably skip all commits from the first NC-188 commit up to the
merge commit.