(NC-188) Rules use group ids instead of names.
The change is basically just returning the :group-id field of the rule
instead of the :group-name (and changing the schema of apply-rule
accordingly).

LOTS OF THINGS ARE STILL BROKEN. The classification and acceptance tests
still mostly error out. If you are bisecting, you should probably skip
all commits from the first NC-188 commit up to the merge commit.