(maint) Simplify uuid? implementation, add tests.
Previously, uuid? would error when passed things that weren't UUIDs or
strings. Change the predicate to just coerce the result of converting
the argument with `->uuid` to a boolean, simplifying the implementation
and increasing safety.

Add tests to make sure that testing and converting things besides UUIDs
or Strings doesn't error.