(NC-188) Return 400 on malformed parent UUIDs.
Now that the parent field is also a UUID, when a group is submitted with
a parent field and its value is not a valid UUID, return a 400 Bad
Request response, with an error message indicating that the parent field
was malformed.

ACCEPTANCE TESTS ARE STILL BROKEN. If you are bisecting, you should
probably skip all commits from the first NC-188 commit until the next
commit, where the acceptance tests get updated and everything passes
again.