(NC-162) Create groups via POST.
Add the capability to create a group by POSTing to the groups collection
endpoint. This is done by reusing the group-resource function, but
passing it nil instead of the UUID from the URI (since there is none).

Since it can now be passed nil, group-resource needs to check if the
uuid argument its given is nil before using it in a few places.

Update the `http/malformed-group?` docstring to describe the validation
it does.