(NC-190) Group's rule field is just the condition.
Change the Group schema to make the :rule key just a RuleCondition,
rather than an entire Rule. Accordingly, enforce that Rules always have
a :group-id by making that key non-optional in the schema. Due to this
change, a couple of preconditions that checked rules for this key can be
removed.

Update all the tests to submit group rules in the new fashion.

Update the group API documentation and README group creation example
accordingly.