(NC-198) Add classification/inheritance-maxima.
This function, when given a map of groups to their ancestors, returns
those groups that are maxima according to the partial ordering defined
by inheritance (where a < b iff a is an ancestor of b). The returned
groups are not less than any other given group by this ordering, meaning
that they are unrelated to all other returned groups.