(NC-198) Add classification/conflicts.
Add a new schema, ClassificationConflict, for the description of
conflicts between classifications, which looks like a Classification
except that the values of all maps are sets of Classification values.

Add classification/conflicts, which takes classifications and returns
a ClassificationConflict map with the conflicts, or nil if there are no
conflicts. Also add some tests for this function.