(NC-197) Add in-memory implementation of Storage.
Add a new namespace, puppetlabs.classifier.storage.memory, that contains
a function `in-memory-storage` which produces an instance of the Storage
protocol that just stores things in an atom in memory. The point of this
is to replace the ad-hoc in memory instances that we've been
constructing when needed in the HTTP tests. It will also be used for the
PermissionedStorage tests.