(maint) Acceptance test instance logs on stderr.
This allows you to seperate the test instance's output from lein test's
output (which goes to stdout) by redirecting one file descriptor or the
other. Without this, it is impossible to figure out what failed in the
acceptance tests because it is intermingled with several thousand lines
of debug log output from the test instance.