How to wire classifier up to a Puppet master
Added a section to the doc describing how to connect the classifier to talk to Puppet
(for environments calling) and how to connect Puppet to talk to the classifier
(as a node terminus). At the end of following my own instructions I was successfully
able to get 500 errors from the classifier to my puppet agent. :)