(NC-222) Encode DateTimes to JSON in ISO8601.
Change the http namespace's liberator defmethods for rendering to JSON
to have them encode joda DateTimes, using the ISO8601 formatter.

These methods rely on kitchensink.json's add-common-json-encoders!
function having been called, in order to register kitchensink's cheshire
encoder for DateTimes, which is what enables with-datetime-encoder to do
anything. When started with TK, this is done as part of the service
initialization, but if you're using the handler by itself (e.g. in the
http tests), it's up to you to call this.