(maint) Fix group-class link bug.
The link between a group and class was only set when creating a group or
adding class parameters to an existing group, so it was not possible to
update a group to link it to a class without setting any class
parameters.

Break out the link-setting into a private `add-group-class-link`
function in postgres.clj, and call it from both `create-group` and
`update-group`.