(maint) Fix class8n/conflicts for > 2 groups.
I thought that the function passed to merge-with was passed all
maps' values at once, but instead it's used like a reduce. This means
that the classification/conflicts function would always erroneously
report conflicts whenever it was passed more than two classifications.

Change the conflicts test to use three classifications for this reason.