(NC-226) Store classification in node check-in.
Add a :classification field to the CheckIn schema for the node's
classification. The field's value is either a Classification or nil if
there were conflicts (so no classification was provided).

Add migrations to add/remove the classification column from the
node_check_ins table.

Update the postgres check-in storage method implementations to encode
& decode the classification as appropriate.

Change http/classify-node to put the classification in the check-in if
there are no conflicts before handing it off to storage.

Update acceptance tests to check for the classification as well.