(NC-228) Add trumps field w/out breaking anything.
Add migrations to add & remove the environment_trumps column from the
groups table.

Add the environment-trumps field to the classification schema (since it
is necessary to be able to determine whether classifications conflict),
and hence transitively to groups.

Since the trumps field is not relevant to the final classification,
create a new ClassificationOutput schema that omits the trumps field,
and adjust the CheckIn schema to change its classification field to
a ClassificationOutput instance.

When calculating classification conflicts, for now the trumps field is
just removed; the desired functionality will be implemented in a later
commit.

Add a whooole lotta :environment-trumps keys to groups in tests so they
conform to the new schema.