(maint) Change explain-rule to take RuleConditions
The explain-rule function doesn't actually need any part of the rule
besides the rule-condition, so it seems silly to have to construct
throwaway information just to get an explanation. I think it took Rules
largely because its name is explain-rule, but Rules are hardly used (and
its output is an ExplainedCondition) so it should not be too surprisng
that it takes a RuleCondition instead.