(maint) Always convert hyphens to underscores.
Use cheshire's key-fn affordance to always convert hyphens to
underscores in JSON keys when serializing, and to go the other way
when deserializing JSON from the client.

This requires extensive changes to tests to use the same key-fn when
serializing data to send to test instances/handlers, since we don't want
to have to use underscores in our test data literals like some sort of
ruby programmer. It also requires changing the group and classification
documentation to rename the 'defined-by' classification conflict field
to 'defined_by'

Add tests to check that the underscore transform is in effect for
liberator responses and middleware-generated error responses.