(NC-227) Namespace cfg files, don't bundle log cfg
Move the .cfg files to 'resources/puppetlabs/classifier`, so they will
not collide with other TK projects' .cfg files when they are all bundled
together in one jar.

Update the jar-exclusions so that the logging config files
(`logback.xml` and `log4j.properties`) are excluded from jars, so that
the end-user may configure logging as they please.