(NC-232) Add class8n/validation-tree-difference.
Add a function in the classification namespace to compute the error
difference between two validation trees. The function removes every
error in the first tree that is also present in the second. The trees
must have the same structure. If the difference is a valid validation
tree (i.e. it contains no more errors), then the errors in the first
tree must be a subset of those in the second.