(maint) Fix bug in group handler.
The handler did not convert the string UUID pulled out of the URI into
an actual UUID when the inherited view of a group was requested, which
would cause spurious 404s. It also did not define a not-found handler,
so it could return a non-JSON error response, which we've promised never
to do.