(NC-240) Add groups->tree to schema ns.
The groups->tree function takes a flat sequential collection of groups
and turns it into a hierarchy iff the collection represents a complete
well-structured group hierarchy. If there are duplicate groups, cycles,
or groups that are not reachable from the root, then an error will be
thrown. If not, then the root's HierarchyNode will be returned.