(NC-240) Add `import-hierarchy` storage function.
Add a new function to the Storage protocol and its postgres
implementation, which takes a flat list of groups, turns it into
a hierarchy, and replaces the extant group hierarchy with this new one,
while creating any classes and class parameters that the hierarchy
refers to if they do not already exist. It returns the entire inserted
hierarchy (i.e. the root HierarchyNode, from which all other groups in
the hierarchy can be reached).