(NC-243) Update PermissionedStorage & Tests
Overhaul the PermissionedStorage wrapper implementation to:
  * Use the final set of RBAC permissions for the NC
  * Respect inherited permissions.
  * Give the permission querying functions a list of ancestors to also
    query permissions for (since permissions inherit).
  * Constrain fields of created groups as necessary to not circuitously
    allow "edits" that the user does not have permission to make.
  * Implicitly grant permissions to allow user to edit classification
    and rules of groups if they have permission to modify the children
    of one of its ancestors.
  * Implement the new `get-group-as-inherited` Storage protocol method
    and scrub inherited values that the user doesn't have permission to
    see.
  * Figure out which parts of a delta are actually changes from the
    current values so as to not trigger spurious edit denials.

Update the PermissionedStorage tests to also use the final RBAC
permissions set and to have the permission-querying functions take
a list of ancestors and query their permissions as well.

Fix a bunch of small bugs in the tests such as missing group fields, bad
permissions maps, missing classes in the memory storage, giving the
wrong argument to storage methods, etc.

Add a test that created children have their environment properly vetted
to match the parent's environment if the user doesn't have permission to
edit the environment of children of the parent.