(maint) Fix two small bugs in schema middleware.
When removing the passing values from the schema exception's schema
field, first check to make sure that the schema field is the same size
as the values; if not, the schema is likely a general sequential schema
and shouldn't be modified.

Cast the schema error field received from the our schema exception data
processing to a seq, so it'll print fully instead of just printing as
a lazy sequence pointer.