(maint) Add http/hyphenate-group-keys utility.
Since we don't wholesale convert underscores to hyphens in requests'
JSON payloads, we have to specifically rename the particular keys in
a group that we expect to be hyphenated. Rather than calling rename-keys
with the same set of keys each time, add a utility to consolidate that
logic in one place.