(NC-241) Add import-hierarchy endpoint.
Add an API endpoint, `/v1/import-hierarchy`, that imports a new
hierarchy wholesale, from a flat array of the groups submitted as JSON
in the request's body. If any of the individual groups are malformed,
then a 400 will be returned to the client with a detailed explanation of
the deformities; if the hierarchy as a whole is malformed, then a 422
will be returned, with an explanation of the deformities and a complete
listing of the groups causing them.