(NC-243) Add `get-group-as-inherited` to Storage
This functionality should be implemented in the Storage layer, rather
than in the HTTP handler. Additionally, since RBAC wraps the Storage
layer and RBAC changes inheritance behavior, this needs to be
implemented in Storage in order for it to be RBAC'd properly.

Add implementations of the Storage method for both the postgres and
in-memory implementations of the Storage protocol.