(NC-243) Non-throwing validate-* Storage methods.
Change `validate-group` to not throw when it finds missing referents in
the tree, but instead return the tree (if there are no errors, it
returns nil), since PermissionedStorage will need to scrub the values of
parameters in the tree that the subject is not permitted to view.

For similar reasons, add `validate-delta` as a Storage method that has
the same semantics as `validate-group`.

Change the http ns's validate-group resource to use `validate-group`
according to its new semantics.

Change a postgres test for `validate-group` to check for the expected
validation tree rather than seeing if an error was thrown.