Change beaker configs to use the classifier role
The beaker configs had used the role 'database', which is meant to specify
the puppetdb role for Puppet Enterprise. This change updates all
instances of the term 'database' throughout the integration tests--both
foss and PE-- to use 'classifier' instead. New configs are included in this
change that install the classifier service on a separate node for PE tests.
The Jenkins matrix should update to those configs once this PR is merged.