(NC-243) Add functional hierarchy utilities.
Add utilities for working with group hierarchy trees using zippers.

`hierarchy-zipper` creates a zipper from a HierarchyNode or
ValidationNode.

`tree->groups` unrolls the tree into a flat list of groups.

`viewable-subtrees` takes a hierarchy node and a function for
determining if a given group id is viewable, and returns a flat list of
the groups that are viewable (subject to the inheritance of
permissions through subtrees).