(maint) Fix two in-memory Storage bugs.
The check to ensure the parent exists when validating a group checked
the wrong thing.

In `get-ancestors`, the loop/recur to pick up all the ancestors always
recurred by getting the parent of the argument to the method, rather
than getting the parent of the current group. This would prevent the
method from finding more than one ancestor.