(NC-244) Add get-groups-ancestors to storage ns.
Add a new function to the storage namespace, `get-groups-ancestors`,
which fetches ancestors for multiple groups without fetching the same
group twice, in order to minimize database requests.

It is in the storage namespace because it is kind of like a Storage
protocol method, but it can be implemented in terms of actual Storage
protocol methods so making it a method would require each implementation
to have to re-define it. It doesn't have many options for optimization
so it doesn't seem worth it to make it a real Storage method.