(NC-245) Add inherited rule to group as-inherited.
Add classification/inherited-rule, which returns the inherited rule of
a group given the inheritance chain starting at that group and ending at
the root.

Use this in http/matching-groups-and-ancestors, which determines the
matching groups according to the inherited rule.

Also use this in all the Storage implementations' get-group-as-inherited
methods to add the inherited rule to the returned group.

Add some tests that getting a group as-inherited includes this inherited
rule.