(NC-248) Detect cycles when root's parent changed.
The check for whether changing a group's parent introduces a cycle in
the hierarchy was not performed for the root group. Perform it for all
groups, and only decline to throw an exception if the cycle is just the
root group pointing to itself (a cycle of length 1 where the group is
the root group).