(maint) Add classifier conf get/set to helper.rb
Add two methods to the beaker helper file for the integration tests,
`get_classifier_configuration` and `set_classifier_configuration`. These
methods allow you to get a hash of the current classifier configuration,
and pass in a hash that will be used as the new classifier
configuration.

These methods depend on knowing the location of the classifier's
configuration file, which was previously hardcoded into the pre_suite
"test" that configured the classifier. To that end, add constants to the
helper.rb file, so they're only hardcoded in one place.

Refactor `integration/pre_suite/95_install_classifier_cert` and
`integration/pe_pre_suite/50_configure_and_start_pe-classifier` to use
the above methods instead of doing it ad-hoc with `create_remote_file`.