(maint) Fix `facts` references in documentation.
A while ago, we changed the key for the facts of a node when submitting
to the classification or classification explanation endpoint (and hence
also how to write rules on facts), but we did not update all of the
documentation to reflect that change. This commit changes all references
to a node's `facts` key to instead talk about the `fact` key.