(maint) Fix environment_trumps update bug.
Change the update-group-field postgres storage function to check for
nilness of the new value, rather than truthiness, when deciding whether
to actually perform the update.

Due to the old truthiness check, it was not possible to update a field
value to be false, which meant that it was impossible to turn off the
environment_trumps flag for a group once it had been turned on.